separate the js libs from the go code
