+++ node/olvm/interpreter/runner/runner.go
@@ -26,26 +26,9 @@ func (runner Runner) exec(callString string) (string, string) {
 	if err != nil {
 		panic(err)
 	}
-
+  sourceCode := getCodeFromJsLibs("onExit")
 	// Set the transaction parameters
-	runner.vm.Run(`
-    var list = context.getUpdateIndexList();
-    var storage = context.getStorage();
-    var transaction = {};
-    for (var i = 0; i< list.length; i ++) {
-      var key = list[i];
-      transaction[key] = storage[key];
-    }
-    transaction.__from__ = __from__;
-    transaction.__olt__ = __olt__;
-    `)
-
-	// Set the results
-	runner.vm.Run(`
-    var out = JSON.stringify(transaction);
-    var ret = JSON.stringify(retValue);
-    `)
-
+	runner.vm.Run(sourceCode)
 	output := ""
 	returnValue := ""
